Move `controller/controllerinstallation` into `controllermanager/controller/controllerregistration/controllerinstallation`
- This code is only executed by `gardener-controller-manager`, so it
  should be in `pkg/controllermanager/controller`
- The code in `pkg/controller` is executed by multiple binaries/Gardener
  components, so this isn't a fit here.
