+++ pkg/controllermanager/controller/controllerregistration/controllerinstallation/add.go
