+++ pkg/controllermanager/controller/controllerregistration/controllerinstallation/seed/add_test.go
@@ -20,7 +20,7 @@ import (
 	gardencorev1 "github.com/gardener/gardener/pkg/apis/core/v1"
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	"github.com/gardener/gardener/pkg/client/kubernetes"
-	"github.com/gardener/gardener/pkg/controller/controllerinstallation"
+	"github.com/gardener/gardener/pkg/controllermanager/controller/controllerregistration/controllerinstallation"
 	. "github.com/gardener/gardener/pkg/controllermanager/controller/controllerregistration/controllerinstallation/seed"
 )
